+++ src/welcome/main.ts
@@ -4,6 +4,7 @@ import 'normalize.css/normalize.css';
 import Welcome from './Welcome.svelte';
 
 new Welcome({
+  // Play transitions on first render
   intro: true,
-  target: document.body,
+  target: document.getElementById('app'),
 });